+++ src/deprecated.jl
@@ -1,0 +1,2 @@
+import Base: merge!
+@deprecate merge!(dest::FitsHeader, other::Union{Pair,FitsCard}) push!(dest, other) false